+++ src/Package/Hector/CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Copy-to-clipboard buttons on debug console SQL queries (raw and with interpolated values)
+
 ## [3.1.1] - 2026-06-10
 
 _No changes in this release._